@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SAR.Apps.Server.Objects
+﻿namespace SAR.Apps.Server.Objects
 {
     public class SetupResponse
     {
