@@ -3,7 +3,7 @@ import React, {Component} from "react";
 import SarService from "../Services/SarService";
 import {Row, Col, Form, Button} from "react-bootstrap";
 import { Link, Redirect } from 'react-router-dom'
-import uuid from 'uuid';
+import {v4 as uuid} from 'uuid';
 import {toast} from "react-toastify";
 
 export default class Character extends Component {
