@@ -1,5 +1,5 @@
 class Environment {
-  static BASE_URL = "http://localhost:8001/api";
+  static BASE_URL = "http://192.168.50.93:8007/api";
 }
 
 export default Environment;
